Escape all special chars as js literals in escape_javascript
This makes sure that all characters that have significance in
html or js parsing is escaped, so that the input string is
correctly represented once parsed as javascript.

It corrects a few issues with the old approach:

1. U+2028 and U+2029 were escaped as html instead of as javascript,
   which meant that you had to use innerHTML or similar html parsing
   to get the correct value.
2. When using on a non safe value, the return value was also unsafe
   and any special html chars which would be html escaped when included
   in the template. Because of this html escaping after the js escape,
   the only way to get the actual value is to interpret it as html.
3. Combining issue 1 and 2, U+2028 and U+2029, would be incorrectly double
   escaped if using escape_javascript on a non safe string.
4. Because the string could contain html escaped characters and single
   or double quotes, it could not safely be used in HTML attributes.
5. The null character is a parse error in html script tags, but was not
   escaped.
6. `\r\n` and `\r` was replaced with `\\n`, which meant that the js escape
   changed the content of the string.

In summary, there was no safe and reliable way to transfer arbitrary text
to a javascript string without parsing it as html on the client side. There
were also some inputs that would not be correctly represented at all.

Transfering a server side string (which might be user provided or
otherwise untrusted) to a javascript variable is useful, and should
be supported.

This implementation makes sure that the given input string is correctly
represented when parsed as javascript, with no change to it's contents.
The documented use case of escaping a safe string and then using it
as html would still be supported.

Since the output is guaranteed to not contain any html special chars,
it could be marked safe if wanted, as a html escape would not change it
at all.

Downsides:

1. Removing the implicit html escape of js escaped values (when using
   non-safe input) could present a security risk for anyone using it
   (that treats the returned js string as safe html).
2. The return value will be a little bit longer and a little less easy to
   read, since it will contain strings like `\u0022` instead of `\"`.